+++ rust/crates/imperialism-core/src/city/production/pipeline.rs
@@ -235,7 +235,7 @@ impl GameState {
                 (port.port_tile == home
                     || port.zone.active_tile == Some(home)
                     || port.zone.target_tile == Some(home))
-                .then(|| OceanZoneId::new(index as usize))
+                .then(|| OceanZoneId::new(index))
             })
             .expect("a city that launches a warship has a port zone for its home tile")
     }
